+++ NEWS.md
@@ -3,6 +3,8 @@
 
 * split out `omop_graph_calc()` for separate calculation of nodes & edges, either a) to enable other data to be joined on (e.g. num records) for visualisation or b) so nodes & edges can be passed to an alternative renderer (as yet unknown)   
 * split out `omop_graph_vis()` to enable  other data to be joined on (e.g. num records) for visualisation
+* `omop_graph()` gets separate args `nodesizevar` and `nodesize` so size can be set by a variable column and/or a sizing param
+ 
 
 # omopcept 0.0.6.0 2024-09-26